Move createCollection into own component collectioncreator.
